fix(stores): guard vector backfill against stale overwrites (ARN-216)
Sequence-monotonic vector reconcile + ADR-0173. Completes GREEN after RED DST.
