+++ crates/temper-server/src/state/projection_backfill.rs
@@ -39,11 +39,13 @@ pub(super) fn transition_table_for(
 /// ADR-0155). Shared by the key and vector backfills so they classify entities the
 /// same way — the distinction is the watermark soundness gate.
 pub(super) enum EntityLoadOutcome {
-    /// Loaded — index it from these fields.
-    Fields(serde_json::Value),
-    /// Definitively skippable: deleted, or a phantom with no events. Correctly NOT
-    /// indexed, and NOT a failure (it must not block the watermark).
-    Skip,
+    /// Loaded — index it from these fields, derived at this journal sequence
+    /// (the ADR-0173 / ARN-216 staleness guard for reconciles).
+    Fields(serde_json::Value, u64),
+    /// Definitively skippable: deleted, or a phantom with no events (sequence
+    /// carried for guarded purges). Correctly NOT indexed, and NOT a failure
+    /// (it must not block the watermark).
+    Skip(u64),
     /// The entity exists (it was enumerated from the durable store) but its current
     /// state could not be loaded — no transition table to replay with, an unreadable
     /// snapshot, or a replay error. Indexing it is impossible, so the type must NOT be
@@ -82,9 +84,9 @@ pub(super) async fn load_entity_current_fields(
     .await
     {
         Err(_) => EntityLoadOutcome::LoadFailed,
-        Ok(state) if state.status == "Deleted" => EntityLoadOutcome::Skip,
-        Ok(state) if state.total_event_count == 0 => EntityLoadOutcome::Skip,
-        Ok(state) => EntityLoadOutcome::Fields(state.fields),
+        Ok(state) if state.status == "Deleted" => EntityLoadOutcome::Skip(state.sequence_nr),
+        Ok(state) if state.total_event_count == 0 => EntityLoadOutcome::Skip(state.sequence_nr),
+        Ok(state) => EntityLoadOutcome::Fields(state.fields, state.sequence_nr),
     }
 }
