+++ crates/temper-server/src/storage/mod.rs
@@ -30,8 +30,6 @@ use temper_store_turso::{
 };
 
 use crate::platform_store::PlatformStore;
-#[cfg(feature = "sim")]
-use crate::platform_store::SimPlatformStore;
 use crate::state::trajectory::{TrajectoryEntry, TrajectorySource};
 
 mod published_artifacts;
@@ -95,6 +93,7 @@ pub trait DynEventStore: Send + Sync {
         entity_type: &'a str,
         entity_id: &'a str,
         vector_rows: &'a [temper_runtime::persistence::EntityVectorRow],
+        as_of_sequence: u64,
     ) -> EventStoreFuture<'a, Result<(), PersistenceError>>;
 
     fn vector_candidates<'a>(
@@ -267,6 +266,7 @@ where
         entity_type: &'a str,
         entity_id: &'a str,
         vector_rows: &'a [temper_runtime::persistence::EntityVectorRow],
+        as_of_sequence: u64,
     ) -> EventStoreFuture<'a, Result<(), PersistenceError>> {
         Box::pin(EventStore::backfill_entity_vectors(
             self,
@@ -274,6 +274,7 @@ where
             entity_type,
             entity_id,
             vector_rows,
+            as_of_sequence,
         ))
     }
 
@@ -539,9 +540,10 @@ impl BoxedEventStore {
         entity_type: &str,
         entity_id: &str,
         vector_rows: &[temper_runtime::persistence::EntityVectorRow],
+        as_of_sequence: u64,
     ) -> Result<(), PersistenceError> {
         self.0
-            .backfill_entity_vectors(tenant, entity_type, entity_id, vector_rows)
+            .backfill_entity_vectors(tenant, entity_type, entity_id, vector_rows, as_of_sequence)
             .await
     }
 
@@ -1146,139 +1148,8 @@ pub trait TursoStoreProvider: Send + Sync {
     async fn ensure_tenant(&self, tenant_id: &str) -> Result<bool, PersistenceError>;
 }
 
-/// Composed storage capabilities selected at boot.
-#[derive(Clone)]
-pub struct StorageStack {
-    pub backend: BackendLabel,
-    pub events: BoxedEventStore,
-    pub postgres_pool: Option<PgPool>,
-    pub turso: Option<Arc<dyn TursoStoreProvider>>,
-    pub platform: Option<Arc<dyn PlatformStore>>,
-    pub policies: Option<Arc<dyn PolicyStore>>,
-    pub query_plane: Option<Arc<dyn QueryPlaneStore>>,
-    pub data_only_create: Option<Arc<dyn DataOnlyCreateStore>>,
-    pub trajectory: Option<Arc<dyn TrajectorySink>>,
-    pub metadata: Option<Arc<dyn MetadataStoreProvider>>,
-}
-
-impl StorageStack {
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        backend: BackendLabel,
-        events: BoxedEventStore,
-        postgres_pool: Option<PgPool>,
-        turso: Option<Arc<dyn TursoStoreProvider>>,
-        platform: Option<Arc<dyn PlatformStore>>,
-        policies: Option<Arc<dyn PolicyStore>>,
-        query_plane: Option<Arc<dyn QueryPlaneStore>>,
-        data_only_create: Option<Arc<dyn DataOnlyCreateStore>>,
-        trajectory: Option<Arc<dyn TrajectorySink>>,
-        metadata: Option<Arc<dyn MetadataStoreProvider>>,
-    ) -> Self {
-        Self {
-            backend,
-            events,
-            postgres_pool,
-            turso,
-            platform,
-            policies,
-            query_plane,
-            data_only_create,
-            trajectory,
-            metadata,
-        }
-    }
-
-    pub fn from_postgres(store: PostgresEventStore) -> Self {
-        let store = Arc::new(store);
-        Self::new(
-            BackendLabel::Postgres,
-            BoxedEventStore::from_arc(store.clone()),
-            Some(store.pool().clone()),
-            None,
-            Some(store.clone() as Arc<dyn PlatformStore>),
-            Some(store.clone() as Arc<dyn PolicyStore>),
-            Some(store.clone() as Arc<dyn QueryPlaneStore>),
-            Some(store.clone() as Arc<dyn DataOnlyCreateStore>),
-            Some(store.clone() as Arc<dyn TrajectorySink>),
-            Some(Arc::new(SingleMetadataStoreProvider::new(store))),
-        )
-    }
-
-    pub fn from_turso(store: TursoEventStore) -> Self {
-        let store = Arc::new(store);
-        Self::new(
-            BackendLabel::Turso,
-            BoxedEventStore::from_arc(store.clone()),
-            None,
-            Some(Arc::new(SingleTursoStoreProvider::new(store.clone()))),
-            Some(store.clone() as Arc<dyn PlatformStore>),
-            Some(store.clone() as Arc<dyn PolicyStore>),
-            Some(store.clone() as Arc<dyn QueryPlaneStore>),
-            None,
-            Some(store.clone() as Arc<dyn TrajectorySink>),
-            Some(Arc::new(SingleMetadataStoreProvider::new(store))),
-        )
-    }
-
-    pub fn from_tenant_router(router: TenantStoreRouter) -> Self {
-        let platform_store = Arc::new(router.platform_store().clone()) as Arc<dyn PlatformStore>;
-        let router = Arc::new(router);
-        Self::new(
-            BackendLabel::TursoRouted,
-            BoxedEventStore::from_arc(router.clone()),
-            None,
-            Some(Arc::new(TenantRoutedTursoStoreProvider::new(
-                router.as_ref().clone(),
-            ))),
-            Some(platform_store),
-            Some(router.clone() as Arc<dyn PolicyStore>),
-            Some(router.clone() as Arc<dyn QueryPlaneStore>),
-            None,
-            Some(router.clone() as Arc<dyn TrajectorySink>),
-            Some(Arc::new(TenantRoutedMetadataStoreProvider::new(
-                router.as_ref().clone(),
-            ))),
-        )
-    }
-
-    pub fn from_redis(store: temper_store_redis::RedisEventStore) -> Self {
-        let store = Arc::new(store);
-        Self::new(
-            BackendLabel::Redis,
-            BoxedEventStore::from_arc(store),
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-    }
-
-    #[cfg(feature = "sim")]
-    pub fn from_sim(
-        store: temper_store_sim::SimEventStore,
-        platform_store: Option<Arc<SimPlatformStore>>,
-    ) -> Self {
-        let store = Arc::new(store);
-        let platform = platform_store.map(|store| store as Arc<dyn PlatformStore>);
-        Self::new(
-            BackendLabel::Sim,
-            BoxedEventStore::from_arc(store),
-            None,
-            None,
-            platform,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-    }
-}
+pub use stack::StorageStack;
+mod stack;
 
 struct SingleMetadataStoreProvider {
     store: Arc<dyn MetadataStore>,
