+++ crates/temper-store-turso/src/store/tests/mod.rs
@@ -125,8 +125,15 @@ async fn vector_index_write_behind_candidates_and_partitioning() {
     assert_eq!(candidates[1].vector, vec![0.0, 1.0]);
 
     // Upsert: re-writing item-a's vector replaces (no duplicate row).
+    // `u64::MAX` forces the reconcile (caller knows the rows are current).
     store
-        .backfill_entity_vectors("t", "Item", "item-a", &[row("embed", "m1", vec![0.5, 0.5])])
+        .backfill_entity_vectors(
+            "t",
+            "Item",
+            "item-a",
+            &[row("embed", "m1", vec![0.5, 0.5])],
+            u64::MAX,
+        )
         .await
         .unwrap();
     let candidates = store
@@ -207,7 +214,7 @@ async fn vector_index_reconcile_purges_on_delete_and_empty_rows() {
 
     // The explicit backfill purge (empty rows) is idempotent.
     store
-        .backfill_entity_vectors("t", "Item", "item-a", &[])
+        .backfill_entity_vectors("t", "Item", "item-a", &[], u64::MAX)
         .await
         .unwrap();
     assert!(
